# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs SWAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Afficher la liste des films",
                "operationId": "868c81d9ba6dfd7ef6c2b31502fae834",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Créé un film",
                "operationId": "92bea0b4038986c2ee2a4b185f912af3",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            }
        },
        "/api/films/{id}": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Afficher un film à l'aide de son id",
                "operationId": "a4df4f825f139541ff446b2eaabf793d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Afficher un film à l'aide de son id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Modifier un film",
                "operationId": "e10f6be5a3c1dd6d923fadad783d5ac1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Films"
                ],
                "summary": "Supprimer un film",
                "operationId": "771c10e4bc6b04dc625ec4fef05b4ab0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            }
        },
...
        "/api/people": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Afficher la liste des personnages",
                "operationId": "9c4a4198e78ac76a924cb30187046a1c",
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des personnages"
                    }
                }
            }
        },
        "/api/people/{id}": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Afficher un personnage à l'aide de son id",
                "operationId": "91a4a00c86a8fb033932abdc59c601c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du personnage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne le personnage ciblé par l'id passé en paramètre"
                    }
                }
            },
            "put": {
                "tags": [
                    "People"
                ],
                "summary": "Modifier un personnage",
                "operationId": "b70c32b8efd74cd5427c2722fa878796",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du personnage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Création d'un personnage",
                "operationId": "5380c0547913de0d85be04a9eb84c614",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du personnage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            },
            "delete": {
                "tags": [
                    "People"
                ],
                "summary": "Suppression d'un personnage",
                "operationId": "2ccbbf08d5e8c0a60a6abdb71e73a7a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du personnage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request"
                    },
                    "200": {
                        "description": "Retourne la liste des films"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}